+++ CHANGELOG.md
@@ -119,6 +119,7 @@ repeat panel from url), fixes #2564
 - Notice to makers/users of custom data sources, there is a minor breaking change in 2.2 that
 require an update to custom data sources for them to work in 2.2. [Read this doc](https://github.com/grafana/grafana/tree/master/docs/sources/datasources/plugin_api.md) for more on the
 data source api change.
+- The duplicate query function used in data source editors is changed, and moveMetricQuery function was renamed
 
 # 2.1.3 (2015-08-24)